refactor: fix code
