+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_API_LINK = https://jsonplaceholder.typicode.com/todos/?_limit=6
