+++ src/actions/actions.ts
@@ -1,6 +1,6 @@
 import { SetStateAction } from "react";
 
-import { ITodo } from "../../types";
+import { ITodo } from "../types";
 
 import {
   ADD_TODO,
@@ -19,21 +19,21 @@ export function addTodo(todo: ITodo) {
   return { type: ADD_TODO, payload: todo };
 }
 
-export function toggleTodo(id: ITodo["id"]) {
+export function toggleTodo(id: number) {
   return { type: TOGGLE_TODO, payload: id };
 }
 
-export function removeTodo(id: ITodo["id"]) {
+export function removeTodo(id: number) {
   return { type: REMOVE_TODO, payload: id };
 }
 
 export function updateTodo(
-  id: ITodo["id"],
+  id: number,
   e: { target: { value: SetStateAction<string> } }
 ) {
   return { type: UPDATE_TODO, payload: id, value: e.target.value };
 }
 
-export function toggleEditing(id: ITodo["id"]) {
+export function toggleEditing(id: number) {
   return { type: IS_EDIT, payload: id };
 }
