+++ src/components/App/style.scss
@@ -1,4 +1,10 @@
 .app {
   max-width: 700px;
   width: 100%;
+  
+  &__warning {
+    font-size: 20px;
+    margin-top: 20px;
+    text-align: center;
+  }
 }
