+++ src/components/MainInput/index.tsx
@@ -1,9 +1,11 @@
 import { SetStateAction, useState } from "react";
 import { useDispatch } from "react-redux";
 
-import { addTodo } from "store/actions/actions";
+import { addTodo } from "actions/actions";
 
 import { ITodo } from "types";
+
+import addTodoBtn from "assets/addTodoBtn.svg"
 
 import "./style.scss";
 
@@ -40,20 +42,7 @@ export const MainInput = () => {
       />
 
       <button type='button' className='mainInput__button' onClick={addBtn}>
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          width='20'
-          height='20'
-          focusable='false'
-          viewBox='0 0 12 12'
-        >
-          <path
-            stroke='#ffffff'
-            strokeLinecap='round'
-            strokeWidth='2'
-            d='M6 2v8m4-4H2'
-          />
-        </svg>
+      <img src={addTodoBtn} alt="Add"/> 
       </button>
     </div>
   );
