+++ src/components/RemoveButton/index.tsx
@@ -1,7 +1,9 @@
 import { useDispatch } from "react-redux";
 
-import { removeTodo } from "store/actions/actions";
+import { removeTodo } from "actions/actions";
 import { IId } from "types";
+
+import removeBtnIcon from "assets/removeBtn.svg"
 
 import "styles.scss";
 
@@ -12,19 +14,7 @@ export const RemoveButton: React.FC<IId> = ({ id }) => {
 
   return (
     <button type='button' className='button' onClick={removeBtn}>
-      <svg
-        xmlns='http://www.w3.org/2000/svg'
-        width='30'
-        height='30'
-        fill='#ffffff'
-        viewBox='0 0 16 16'
-      >
-        <path d='M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6z' />
-        <path
-          fillRule='evenodd'
-          d='M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118zM2.5 3V2h11v1h-11z'
-        />
-      </svg>
+      <img src={removeBtnIcon} alt="Remove"/>
     </button>
   );
 };
