+++ src/components/SaveButton/index.tsx
@@ -1,43 +1,13 @@
 import { IMouseEvent } from "types";
 
+import saveBtn from "assets/saveBtn.svg"
+
 import "styles.scss";
 
 export const SaveButton: React.FC<IMouseEvent> = ({ onToggleIsEditing }) => {
   return (
     <button type='button' className='button orange' onClick={onToggleIsEditing}>
-      <svg
-        width='30px'
-        height='30px'
-        viewBox='0 0 18 18'
-        version='1.1'
-        xmlns='http://www.w3.org/2000/svg'
-      >
-        <g
-          id='Icons'
-          stroke='none'
-          strokeWidth='1'
-          fill='none'
-          fillRule='evenodd'
-        >
-          <g id='Rounded' transform='translate(-443.000000, -1529.000000)'>
-            <g id='Content' transform='translate(100.000000, 1428.000000)'>
-              <g
-                id='-Round-/-Content-/-save_alt'
-                transform='translate(340.000000, 98.000000)'
-              >
-                <g>
-                  <polygon id='Path' points='0 0 24 0 24 24 0 24'></polygon>
-                  <path
-                    fill='#ffffff'
-                    d='M19,13 L19,18 C19,18.55 18.55,19 18,19 L6,19 C5.45,19 5,18.55 5,18 L5,13 C5,12.45 4.55,12 4,12 C3.45,12 3,12.45 3,13 L3,19 C3,20.1 3.9,21 5,21 L19,21 C20.1,21 21,20.1 21,19 L21,13 C21,12.45 20.55,12 20,12 C19.45,12 19,12.45 19,13 Z M13,12.67 L14.88,10.79 C15.27,10.4 15.9,10.4 16.29,10.79 C16.68,11.18 16.68,11.81 16.29,12.2 L12.7,15.79 C12.31,16.18 11.68,16.18 11.29,15.79 L7.7,12.2 C7.31,11.81 7.31,11.18 7.7,10.79 C8.09,10.4 8.72,10.4 9.11,10.79 L11,12.67 L11,4 C11,3.45 11.45,3 12,3 C12.55,3 13,3.45 13,4 L13,12.67 Z'
-                    id='🔹Icon-Color'
-                  ></path>
-                </g>
-              </g>
-            </g>
-          </g>
-        </g>
-      </svg>
+      <img src={saveBtn} alt="Save"/>
     </button>
   );
 };
