+++ src/components/TodoList/index.tsx
@@ -1,12 +1,11 @@
-import TodoItem from "components/TodoItem";
+import TodoItem from "components/TodoItem/container";
 
 import { ITodoList, ITodo } from "types";
 
 export const TodoList: React.FC<ITodoList> = ({ todos }) => {
   return (
     <ul>
-      {todos &&
-        todos.map((item: ITodo) => {
+      {todos?.map((item: ITodo) => {
           return <TodoItem key={item.id} todo={item} />;
         })}
     </ul>
