+++ src/constants/links.ts
@@ -1,0 +1,1 @@
+export const apiLink: RequestInfo = process.env.REACT_APP_API_LINK!;
